fix: 디바이스 등록 로직 수정
