+++ src/app/(device-required)/device/_components/_templates/AddDeviceStep2.tsx
@@ -1,16 +1,11 @@
 "use client";
 
-import { useRouter } from "next/navigation";
-import { useEffect, useState } from "react";
 import ResponsiveButton from "@/components/common/Button/ResponsiveButton";
 import { Form } from "@/components/common/Form";
 import LabeledInput from "@/components/common/LabeledInput";
-import { setEncryptedItem } from "@/lib/auth/secureStorage";
 import cn from "@/lib/utils";
-import makeDeviceName from "@/utils/makeDeviceName";
-import { deviceQueries } from "../../_queries/useDeviceInfo";
+import Spinner from "@/components/common/Spinner";
 import useStep2Form from "../../_hooks/useStep2Form";
-import { TypeDeviceStep2Form } from "../../_schema/device.schema";
 
 type DevicePurpose = "HALL" | "POS";
 
@@ -25,69 +20,11 @@ interface IProps {
   storeName: string;
 }
 
-export default function AddDeviceStep2({
-  storeId,
-  storeName,
-  phoneNumber,
-}: IProps) {
-  const navigate = useRouter();
-  const [purpose, setPurpose] = useState<DevicePurpose>("HALL");
-
-  const {
-    form: { form, watch, handleSubmit },
-  } = useStep2Form();
-
-  useEffect(() => {
-    form.setValue("deviceName", makeDeviceName(purpose));
-    // eslint-disable-next-line
-  }, [purpose]);
-
-  const { mutate } = deviceQueries.useAddDevice();
-
-  const submitHandler = (data: TypeDeviceStep2Form) => {
-    const submitData = {
-      phoneNumber: phoneNumber.replaceAll("-", ""),
-      storeId,
-      name: data.deviceName,
-      purpose,
-      tableNo: 1,
-      paymentType: "POSTPAID" as DevicePayment,
-    };
-
-    mutate(submitData, {
-      onSuccess: async (returnData) => {
-        const { deviceId, secretKey } = returnData;
-        await setEncryptedItem({
-          key: "@deviceInfo",
-          value: {
-            deviceId,
-            storeId: submitData.storeId,
-            storeName,
-            name: submitData.name,
-            purpose: submitData.purpose,
-          },
-          deviceId: String(deviceId),
-          storeId: submitData.storeId,
-        });
-
-        await setEncryptedItem({
-          key: "@secretKey",
-          value: secretKey,
-          deviceId: String(deviceId),
-          storeId: submitData.storeId,
-        });
-        localStorage.setItem(
-          "@meta",
-          JSON.stringify({
-            deviceId,
-            storeId: submitData.storeId,
-          })
-        );
-
-        navigate.replace(purpose === "HALL" ? "/hall" : "/pos");
-      },
+export default function AddDeviceStep2({ ...props }: IProps) {
+  const { form, submitHandler, purpose, setPurpose, isSubmitting } =
+    useStep2Form({
+      ...props,
     });
-  };
 
   return (
     <div className="flex flex-col gap-6">
@@ -111,13 +48,14 @@ export default function AddDeviceStep2({
       <Form {...form}>
         <form
           className="flex flex-col gap-3 lg:gap-8"
-          onSubmit={handleSubmit(submitHandler)}
+          onSubmit={form.handleSubmit(submitHandler)}
         >
           <LabeledInput
             form={form}
             label="기기 이름"
             name="deviceName"
             placeholder="기기 이름을 입력해주세요."
+            readOnly={isSubmitting}
           />
           <ResponsiveButton
             type="submit"
@@ -126,9 +64,9 @@ export default function AddDeviceStep2({
               md: { buttonSize: "sm", className: "w-full mt-3" },
               lg: { buttonSize: "lg", className: "w-full" },
             }}
-            disabled={!watch("deviceName")}
+            disabled={!form.watch("deviceName")}
           >
-            등록하기
+            {isSubmitting ? <Spinner /> : "등록하기"}
           </ResponsiveButton>
         </form>
       </Form>
